Refactor: Rename Appwrite to AppwriteService and update imports
This commit renames the `Appwrite` object to `AppwriteService` to better reflect its purpose as a service. It also updates all relevant import statements and usages of the object across the codebase.
