+++ app/src/main/java/com/example/infiltrados/InfiltradosApplication.kt
@@ -1,11 +1,11 @@
 package com.example.infiltrados
 
 import android.app.Application
-import com.example.infiltrados.backend.Appwrite
+import com.example.infiltrados.services.AppwriteService
 
 class InfiltradosApplication : Application() {
     override fun onCreate() {
         super.onCreate()
-        Appwrite.init(applicationContext)
+        AppwriteService.init(applicationContext)
     }
 }
