+++ app/src/main/java/com/example/infiltrados/ui/main/multiplayer/MultiplayerGameViewModel.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.infiltrados.backend.GameRecord
+import com.example.infiltrados.services.GameRecord
 import com.example.infiltrados.services.MultiplayerGameManager
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
